Edit Event functionality added

+++ app/src/main/java/com/sid1804492/bottomnavtest/database/TeacherPlannerDao.kt
@@ -85,7 +85,7 @@ interface TeacherPlannerDao {
     suspend fun getClass(key: Long): SchoolClass?
 
     @Query("SELECT * FROM event_table WHERE EventId = :key")
-    suspend fun getEvent(key: Int): Event?
+    suspend fun getEvent(key: Long): Event?
 
     @Query("SELECT * FROM classes_table")
     fun getAllClasses(): LiveData<List<SchoolClass>>
